Install benchmarks using "ubuntu", the non-root user
Currently, we use sudo to run the scripts that download and compile
the benchmarks. This results in the benchmarks being owned by the
root user.

This commit removes sudo when executing the run scripts, and only
uses sudo where appropriate. As a result, the benchmarks are owned
by the default non-root user, "ubuntu".

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>